- added gradle wrapper - renamed module for kafka experiments to kafka - fixed kotlin conventions plugin
