+++ buildSrc/build.gradle
@@ -4,7 +4,6 @@ plugins {
 }
 
 repositories {
-    // Use the plugin portal to apply community plugins in convention plugins.
     mavenCentral()
 }
