+++ buildSrc/src/main/groovy/ayansen.playground.kotlin-common-conventions.gradle
@@ -10,23 +10,16 @@ repositories {
 }
 
 dependencies {
-    constraints {
-        implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
-    }
-
     // Align versions of all Kotlin components
     implementation platform('org.jetbrains.kotlin:kotlin-bom')
 
     // Use the Kotlin JDK 8 standard library.
     implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
 
-    // Align versions of all Kotlin components
-    implementation platform('org.jetbrains.kotlin:kotlin-bom')
-
     // Align tests to junit5 of all Kotlin components
-    testImplementation ('org.junit.jupiter:junit-jupiter-api:5.9.0')
-    testImplementation ('org.junit.jupiter:junit-jupiter-engine:5.9.0')
-    testImplementation platform('org.jetbrains.kotlin:kotlin-test-junit5')
+    testImplementation ('org.junit.jupiter:junit-jupiter-api')
+    testImplementation ('org.junit.jupiter:junit-jupiter-engine')
+    testImplementation ('org.jetbrains.kotlin:kotlin-test-junit5')
 }
 
 tasks.named('test') {
