+++ kafka/README.md
@@ -1,0 +1,6 @@
+# kafka
+This is a repository for trying out different features around kafka and kafka-streams.
+
+## Tests
+ - Ordering of events across partitions - [Test](./src/test/kotlin/ayansen/playground/kafka/mirror/EventOrderingIT.kt) to validate kakfa events are ordered in a partition 
+
