+++ kafka/build.gradle
