+++ kafka/src/main/kotlin/ayansen/playground/kafka/mirror/KafkaMirror.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ayansen.programming.kafka.experiments
+package ayansen.playground.kafka.mirror
 
 import java.util.Properties
 import kotlin.concurrent.thread
